Retrieve all valid plotting symbols for the plotly library

Retrieve all valid plotting symbols for the plotly library

Usage

Symbol_List()

Value

A vector of all the valid plotting symbols used in the plot_ly library.

Examples

Symbol_List()
#>   [1] "0"                           "circle"                     
#>   [3] "100"                         "circle-open"                
#>   [5] "200"                         "circle-dot"                 
#>   [7] "300"                         "circle-open-dot"            
#>   [9] "1"                           "square"                     
#>  [11] "101"                         "square-open"                
#>  [13] "201"                         "square-dot"                 
#>  [15] "301"                         "square-open-dot"            
#>  [17] "2"                           "diamond"                    
#>  [19] "102"                         "diamond-open"               
#>  [21] "202"                         "diamond-dot"                
#>  [23] "302"                         "diamond-open-dot"           
#>  [25] "3"                           "cross"                      
#>  [27] "103"                         "cross-open"                 
#>  [29] "203"                         "cross-dot"                  
#>  [31] "303"                         "cross-open-dot"             
#>  [33] "4"                           "x"                          
#>  [35] "104"                         "x-open"                     
#>  [37] "204"                         "x-dot"                      
#>  [39] "304"                         "x-open-dot"                 
#>  [41] "5"                           "triangle-up"                
#>  [43] "105"                         "triangle-up-open"           
#>  [45] "205"                         "triangle-up-dot"            
#>  [47] "305"                         "triangle-up-open-dot"       
#>  [49] "6"                           "triangle-down"              
#>  [51] "106"                         "triangle-down-open"         
#>  [53] "206"                         "triangle-down-dot"          
#>  [55] "306"                         "triangle-down-open-dot"     
#>  [57] "7"                           "triangle-left"              
#>  [59] "107"                         "triangle-left-open"         
#>  [61] "207"                         "triangle-left-dot"          
#>  [63] "307"                         "triangle-left-open-dot"     
#>  [65] "8"                           "triangle-right"             
#>  [67] "108"                         "triangle-right-open"        
#>  [69] "208"                         "triangle-right-dot"         
#>  [71] "308"                         "triangle-right-open-dot"    
#>  [73] "9"                           "triangle-ne"                
#>  [75] "109"                         "triangle-ne-open"           
#>  [77] "209"                         "triangle-ne-dot"            
#>  [79] "309"                         "triangle-ne-open-dot"       
#>  [81] "10"                          "triangle-se"                
#>  [83] "110"                         "triangle-se-open"           
#>  [85] "210"                         "triangle-se-dot"            
#>  [87] "310"                         "triangle-se-open-dot"       
#>  [89] "11"                          "triangle-sw"                
#>  [91] "111"                         "triangle-sw-open"           
#>  [93] "211"                         "triangle-sw-dot"            
#>  [95] "311"                         "triangle-sw-open-dot"       
#>  [97] "12"                          "triangle-nw"                
#>  [99] "112"                         "triangle-nw-open"           
#> [101] "212"                         "triangle-nw-dot"            
#> [103] "312"                         "triangle-nw-open-dot"       
#> [105] "13"                          "pentagon"                   
#> [107] "113"                         "pentagon-open"              
#> [109] "213"                         "pentagon-dot"               
#> [111] "313"                         "pentagon-open-dot"          
#> [113] "14"                          "hexagon"                    
#> [115] "114"                         "hexagon-open"               
#> [117] "214"                         "hexagon-dot"                
#> [119] "314"                         "hexagon-open-dot"           
#> [121] "15"                          "hexagon2"                   
#> [123] "115"                         "hexagon2-open"              
#> [125] "215"                         "hexagon2-dot"               
#> [127] "315"                         "hexagon2-open-dot"          
#> [129] "16"                          "octagon"                    
#> [131] "116"                         "octagon-open"               
#> [133] "216"                         "octagon-dot"                
#> [135] "316"                         "octagon-open-dot"           
#> [137] "17"                          "star"                       
#> [139] "117"                         "star-open"                  
#> [141] "217"                         "star-dot"                   
#> [143] "317"                         "star-open-dot"              
#> [145] "18"                          "hexagram"                   
#> [147] "118"                         "hexagram-open"              
#> [149] "218"                         "hexagram-dot"               
#> [151] "318"                         "hexagram-open-dot"          
#> [153] "19"                          "star-triangle-up"           
#> [155] "119"                         "star-triangle-up-open"      
#> [157] "219"                         "star-triangle-up-dot"       
#> [159] "319"                         "star-triangle-up-open-dot"  
#> [161] "20"                          "star-triangle-down"         
#> [163] "120"                         "star-triangle-down-open"    
#> [165] "220"                         "star-triangle-down-dot"     
#> [167] "320"                         "star-triangle-down-open-dot"
#> [169] "21"                          "star-square"                
#> [171] "121"                         "star-square-open"           
#> [173] "221"                         "star-square-dot"            
#> [175] "321"                         "star-square-open-dot"       
#> [177] "22"                          "star-diamond"               
#> [179] "122"                         "star-diamond-open"          
#> [181] "222"                         "star-diamond-dot"           
#> [183] "322"                         "star-diamond-open-dot"      
#> [185] "23"                          "diamond-tall"               
#> [187] "123"                         "diamond-tall-open"          
#> [189] "223"                         "diamond-tall-dot"           
#> [191] "323"                         "diamond-tall-open-dot"      
#> [193] "24"                          "diamond-wide"               
#> [195] "124"                         "diamond-wide-open"          
#> [197] "224"                         "diamond-wide-dot"           
#> [199] "324"                         "diamond-wide-open-dot"      
#> [201] "25"                          "hourglass"                  
#> [203] "125"                         "hourglass-open"             
#> [205] "26"                          "bowtie"                     
#> [207] "126"                         "bowtie-open"                
#> [209] "27"                          "circle-cross"               
#> [211] "127"                         "circle-cross-open"          
#> [213] "28"                          "circle-x"                   
#> [215] "128"                         "circle-x-open"              
#> [217] "29"                          "square-cross"               
#> [219] "129"                         "square-cross-open"          
#> [221] "30"                          "square-x"                   
#> [223] "130"                         "square-x-open"              
#> [225] "31"                          "diamond-cross"              
#> [227] "131"                         "diamond-cross-open"         
#> [229] "32"                          "diamond-x"                  
#> [231] "132"                         "diamond-x-open"             
#> [233] "33"                          "cross-thin"                 
#> [235] "133"                         "cross-thin-open"            
#> [237] "34"                          "x-thin"                     
#> [239] "134"                         "x-thin-open"                
#> [241] "35"                          "asterisk"                   
#> [243] "135"                         "asterisk-open"              
#> [245] "36"                          "hash"                       
#> [247] "136"                         "hash-open"                  
#> [249] "236"                         "hash-dot"                   
#> [251] "336"                         "hash-open-dot"              
#> [253] "37"                          "y-up"                       
#> [255] "137"                         "y-up-open"                  
#> [257] "38"                          "y-down"                     
#> [259] "138"                         "y-down-open"                
#> [261] "39"                          "y-left"                     
#> [263] "139"                         "y-left-open"                
#> [265] "40"                          "y-right"                    
#> [267] "140"                         "y-right-open"               
#> [269] "41"                          "line-ew"                    
#> [271] "141"                         "line-ew-open"               
#> [273] "42"                          "line-ns"                    
#> [275] "142"                         "line-ns-open"               
#> [277] "43"                          "line-ne"                    
#> [279] "143"                         "line-ne-open"               
#> [281] "44"                          "line-nw"                    
#> [283] "144"                         "line-nw-open"               
#> [285] "45"                          "arrow-up"                   
#> [287] "145"                         "arrow-up-open"              
#> [289] "46"                          "arrow-down"                 
#> [291] "146"                         "arrow-down-open"            
#> [293] "47"                          "arrow-left"                 
#> [295] "147"                         "arrow-left-open"            
#> [297] "48"                          "arrow-right"                
#> [299] "148"                         "arrow-right-open"           
#> [301] "49"                          "arrow-bar-up"               
#> [303] "149"                         "arrow-bar-up-open"          
#> [305] "50"                          "arrow-bar-down"             
#> [307] "150"                         "arrow-bar-down-open"        
#> [309] "51"                          "arrow-bar-left"             
#> [311] "151"                         "arrow-bar-left-open"        
#> [313] "52"                          "arrow-bar-right"            
#> [315] "152"                         "arrow-bar-right-open"       
#> [317] "53"                          "arrow"                      
#> [319] "153"                         "arrow-open"                 
#> [321] "54"                          "arrow-wide"                 
#> [323] "154"                         "arrow-wide-open"